requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
twilio
python-multipart
